backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
better-auth==0.0.1-beta.9
python-dotenv==1.0.1
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.6.1
qdrant-client==1.12.1
openai==1.51.2
torch==2.5.0
transformers==4.45.2
sentence-transformers==3.1.1
langdetect==1.0.9
googletrans==4.0.0rc1